migrate eslint
