+++ package.json
@@ -6,58 +6,58 @@
   "dependencies": {
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.1",
-    "@hookform/resolvers": "^5.4.0",
+    "@hookform/resolvers": "^5.9.1",
     "@mui/icons-material": "^7.3.11",
     "@mui/material": "^7.3.11",
-    "@supabase/supabase-js": "^2.108.1",
-    "@tanstack/react-query": "^5.101.0",
-    "@vis.gl/react-google-maps": "^1.8.3",
+    "@supabase/supabase-js": "^2.116.0",
+    "@tanstack/react-query": "^5.102.8",
+    "@vis.gl/react-google-maps": "^1.10.0",
     "date-fns": "^4.4.0",
-    "posthog-js": "^1.383.3",
-    "react": "^19.2.7",
-    "react-dom": "^19.2.7",
+    "posthog-js": "^1.428.7",
+    "react": "^19.2.8",
+    "react-dom": "^19.2.8",
     "react-dropzone": "^14.4.1",
     "react-hook-form": "7.69.0",
-    "react-router": "^7.17.0",
+    "react-router": "^7.18.3",
     "react-toastify": "^11.1.0",
-    "sass": "^1.100.0",
-    "zod": "^4.4.3"
+    "sass": "^1.104.0",
+    "zod": "^4.5.4"
   },
   "pre-commit": [
     "lint"
   ],
   "devDependencies": {
     "@chromatic-com/storybook": "^4.1.3",
-    "@eslint/js": "^9.39.4",
+    "@eslint/js": "^10.0.1",
     "@hookform/devtools": "^4.4.0",
-    "@storybook/addon-docs": "^10.4.3",
-    "@storybook/addon-links": "^10.4.3",
-    "@storybook/addon-onboarding": "^10.4.3",
-    "@storybook/react-vite": "^10.4.3",
-    "@types/google.maps": "^3.65.1",
-    "@types/node": "^25.9.2",
-    "@types/react": "^19.2.17",
-    "@types/react-dom": "^19.2.3",
+    "@storybook/addon-docs": "^10.6.0",
+    "@storybook/addon-links": "^10.6.0",
+    "@storybook/addon-onboarding": "^10.6.0",
+    "@storybook/react-vite": "^10.6.0",
+    "@types/google.maps": "^3.66.2",
+    "@types/node": "^25.9.5",
+    "@types/react": "^19.2.18",
+    "@types/react-dom": "^19.2.7",
     "@vitejs/plugin-react": "^5.2.0",
-    "cypress": "^15.17.0",
-    "cypress-vite": "^1.10.0",
-    "eslint": "^9.39.4",
+    "cypress": "^15.21.1",
+    "cypress-vite": "^1.10.2",
+    "eslint": "^10.10.0",
     "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-cypress": "^5.3.0",
+    "eslint-plugin-cypress": "^7.0.2",
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^7.1.1",
-    "eslint-plugin-react-refresh": "^0.4.26",
-    "eslint-plugin-storybook": "^10.4.3",
+    "eslint-plugin-react-refresh": "^0.5.6",
+    "eslint-plugin-storybook": "^10.6.0",
     "globals": "^16.5.0",
     "pre-commit": "^1.2.2",
     "prettier": "^2.8.8",
-    "storybook": "^10.4.3",
-    "supabase": "^2.105.0",
+    "storybook": "^10.6.0",
+    "supabase": "^2.117.0",
     "typescript": "^5.9.3",
-    "typescript-eslint": "^8.61.0",
-    "vite": "^7.3.5",
+    "typescript-eslint": "^8.70.0",
+    "vite": "^7.3.6",
     "vite-plugin-eslint": "^1.8.1",
     "vite-tsconfig-paths": "^6.1.1"
   },
